Separate MessagesViewController functionality into files
