+++ Sources/Controllers/MessagesViewController.swift
@@ -1,4 +1,3 @@
-
 /*
  MIT License
 
@@ -58,14 +57,14 @@ open class MessagesViewController: UIViewController {
     open override var shouldAutorotate: Bool {
         return false
     }
-    
-    /// Indicated selected indexPath when handle menu action
-    fileprivate var selectedIndexPathForMenu: IndexPath?
-    
+
     /// A Boolean value used to determine if `viewDidLayoutSubviews()` has been called.
     private var isFirstLayout: Bool = true
     
-    private var messageCollectionViewBottomInset: CGFloat = 0 {
+    /// Indicated selected indexPath when handle menu action
+    var selectedIndexPathForMenu: IndexPath?
+
+    var messageCollectionViewBottomInset: CGFloat = 0 {
         didSet {
             messagesCollectionView.contentInset.bottom = messageCollectionViewBottomInset
             messagesCollectionView.scrollIndicatorInsets.bottom = messageCollectionViewBottomInset
@@ -76,19 +75,12 @@ open class MessagesViewController: UIViewController {
 
     open override func viewDidLoad() {
         super.viewDidLoad()
-
-        extendedLayoutIncludesOpaqueBars = true
-        automaticallyAdjustsScrollViewInsets = false
-        view.backgroundColor = .white
-        messagesCollectionView.keyboardDismissMode = .interactive
-        messagesCollectionView.alwaysBounceVertical = true
-        
+        setupDefaults()
         setupSubviews()
         setupConstraints()
         registerReusableViews()
         setupDelegates()
         addMenuControllerObservers()
-
     }
 
     open override func viewDidLayoutSubviews() {
@@ -109,15 +101,28 @@ open class MessagesViewController: UIViewController {
 
     // MARK: - Methods [Private]
 
+    /// Sets the default values for the MessagesViewController
+    private func setupDefaults() {
+        extendedLayoutIncludesOpaqueBars = true
+        automaticallyAdjustsScrollViewInsets = false
+        view.backgroundColor = .white
+        messagesCollectionView.keyboardDismissMode = .interactive
+        messagesCollectionView.alwaysBounceVertical = true
+    }
+
     /// Sets the delegate and dataSource of the messagesCollectionView property.
     private func setupDelegates() {
         messagesCollectionView.delegate = self
         messagesCollectionView.dataSource = self
     }
 
+    /// Adds the messagesCollectionView to the controllers root view.
+    private func setupSubviews() {
+        view.addSubview(messagesCollectionView)
+    }
+
     /// Registers all cells and supplementary views of the messagesCollectionView property.
     private func registerReusableViews() {
-
         messagesCollectionView.register(TextMessageCell.self)
         messagesCollectionView.register(MediaMessageCell.self)
         messagesCollectionView.register(LocationMessageCell.self)
@@ -125,12 +130,6 @@ open class MessagesViewController: UIViewController {
         messagesCollectionView.register(MessageFooterView.self, forSupplementaryViewOfKind: UICollectionElementKindSectionFooter)
         messagesCollectionView.register(MessageHeaderView.self, forSupplementaryViewOfKind: UICollectionElementKindSectionHeader)
         messagesCollectionView.register(MessageDateHeaderView.self, forSupplementaryViewOfKind: UICollectionElementKindSectionHeader)
-
-    }
-
-    /// Adds the messagesCollectionView to the controllers root view.
-    private func setupSubviews() {
-        view.addSubview(messagesCollectionView)
     }
 
     /// Sets the constraints of the `MessagesCollectionView`.
@@ -149,286 +148,5 @@ open class MessagesViewController: UIViewController {
             NSLayoutConstraint.activate([top, bottom, trailing, leading])
         }
         adjustScrollViewInset()
-    }
-    
-    @objc
-    private func adjustScrollViewInset() {
-        if #available(iOS 11.0, *) {
-            // No need to add to the top contentInset
-        } else {
-            let navigationBarInset = navigationController?.navigationBar.frame.height ?? 0
-            let statusBarInset: CGFloat = UIApplication.shared.isStatusBarHidden ? 0 : 20
-            let topInset = navigationBarInset + statusBarInset
-            messagesCollectionView.contentInset.top = topInset
-            messagesCollectionView.scrollIndicatorInsets.top = topInset
-        }
-    }
-}
-
-// MARK: - UICollectionViewDelegate & UICollectionViewDelegateFlowLayout Conformance
-
-extension MessagesViewController: UICollectionViewDelegateFlowLayout {
-
-    open func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        guard let messagesFlowLayout = collectionViewLayout as? MessagesCollectionViewFlowLayout else { return .zero }
-        return messagesFlowLayout.sizeForItem(at: indexPath)
-    }
-    
-    open func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
-        guard let messagesCollectionView = collectionView as? MessagesCollectionView else { return .zero }
-        guard let messagesDataSource = messagesCollectionView.messagesDataSource else { return .zero }
-        guard let messagesLayoutDelegate = messagesCollectionView.messagesLayoutDelegate else { return .zero }
-        // Could pose a problem if subclass behaviors allows more than one item per section
-        let indexPath = IndexPath(item: 0, section: section)
-        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-        return messagesLayoutDelegate.headerViewSize(for: message, at: indexPath, in: messagesCollectionView)
-    }
-    
-    open func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForFooterInSection section: Int) -> CGSize {
-        guard let messagesCollectionView = collectionView as? MessagesCollectionView else { return .zero }
-        guard let messagesDataSource = messagesCollectionView.messagesDataSource else { return .zero }
-        guard let messagesLayoutDelegate = messagesCollectionView.messagesLayoutDelegate else { return .zero }
-        // Could pose a problem if subclass behaviors allows more than one item per section
-        let indexPath = IndexPath(item: 0, section: section)
-        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-        return messagesLayoutDelegate.footerViewSize(for: message, at: indexPath, in: messagesCollectionView)
-    }
-    
-    public func collectionView(_ collectionView: UICollectionView, shouldShowMenuForItemAt indexPath: IndexPath) -> Bool {
-        guard let messagesDataSource = messagesCollectionView.messagesDataSource else { return false }
-        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-        
-        switch message.data {
-        case .text, .attributedText, .emoji, .photo:
-            selectedIndexPathForMenu = indexPath
-            return true
-        default:
-            return false
-        }
-    }
-    
-    public func collectionView(_ collectionView: UICollectionView, canPerformAction action: Selector, forItemAt indexPath: IndexPath, withSender sender: Any?) -> Bool {
-        return (action == NSSelectorFromString("copy:"))
-    }
-    
-    public func collectionView(_ collectionView: UICollectionView, performAction action: Selector, forItemAt indexPath: IndexPath, withSender sender: Any?) {
-        guard let messagesDataSource = messagesCollectionView.messagesDataSource else { fatalError("Please set messagesDataSource") }
-        let pasteBoard = UIPasteboard.general
-        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-        
-        switch message.data {
-        case .text(let text), .emoji(let text):
-            pasteBoard.string = text
-        case .attributedText(let attributedText):
-            pasteBoard.string = attributedText.string
-        case .photo(let image):
-            pasteBoard.image = image
-        default:
-            break
-        }
-    }
-}
-
-// MARK: - UICollectionViewDataSource Conformance
-
-extension MessagesViewController: UICollectionViewDataSource {
-
-    open func numberOfSections(in collectionView: UICollectionView) -> Int {
-        guard let collectionView = collectionView as? MessagesCollectionView else { return 0 }
-
-        // Each message is its own section
-        return collectionView.messagesDataSource?.numberOfMessages(in: collectionView) ?? 0
-    }
-
-    open func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        guard let collectionView = collectionView as? MessagesCollectionView else { return 0 }
-
-        let messageCount = collectionView.messagesDataSource?.numberOfMessages(in: collectionView) ?? 0
-        // There will only ever be 1 message per section
-        return messageCount > 0 ? 1 : 0
-
-    }
-
-    open func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-
-        guard let messagesCollectionView = collectionView as? MessagesCollectionView else {
-            fatalError("Managed collectionView: \(collectionView.debugDescription) is not a MessagesCollectionView.")
-        }
-
-        guard let messagesDataSource = messagesCollectionView.messagesDataSource else {
-            fatalError("MessagesDataSource has not been set.")
-        }
-
-        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-
-        switch message.data {
-        case .text, .attributedText, .emoji:
-            let cell = messagesCollectionView.dequeueReusableCell(TextMessageCell.self, for: indexPath)
-            cell.configure(with: message, at: indexPath, and: messagesCollectionView)
-            return cell
-        case .photo, .video:
-    	    let cell = messagesCollectionView.dequeueReusableCell(MediaMessageCell.self, for: indexPath)
-            cell.configure(with: message, at: indexPath, and: messagesCollectionView)
-            return cell
-        case .location:
-    	    let cell = messagesCollectionView.dequeueReusableCell(LocationMessageCell.self, for: indexPath)
-            cell.configure(with: message, at: indexPath, and: messagesCollectionView)
-            return cell
-        }
-    }
-
-    open func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-
-        guard let messagesCollectionView = collectionView as? MessagesCollectionView else {
-            fatalError("Managed collectionView: \(collectionView.debugDescription) is not a MessagesCollectionView.")
-        }
-
-        guard let dataSource = messagesCollectionView.messagesDataSource else {
-            fatalError("MessagesDataSource has not been set.")
-        }
-
-        guard let displayDelegate = messagesCollectionView.messagesDisplayDelegate else {
-            fatalError("MessagesDisplayDelegate has not been set.")
-        }
-
-        let message = dataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-
-        switch kind {
-        case UICollectionElementKindSectionHeader:
-            return displayDelegate.messageHeaderView(for: message, at: indexPath, in: messagesCollectionView)
-        case UICollectionElementKindSectionFooter:
-            return displayDelegate.messageFooterView(for: message, at: indexPath, in: messagesCollectionView)
-        default:
-            fatalError("Unrecognized element of kind: \(kind)")
-        }
-    }
-}
-
-// MARK: - Keyboard Handling
-
-fileprivate extension MessagesViewController {
-
-    func addKeyboardObservers() {
-        NotificationCenter.default.addObserver(self, selector: #selector(handleKeyboardDidChangeState), name: .UIKeyboardWillChangeFrame, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(handleTextViewDidBeginEditing), name: .UITextViewTextDidBeginEditing, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(adjustScrollViewInset), name: .UIDeviceOrientationDidChange, object: nil)
-    }
-
-    func removeKeyboardObservers() {
-        NotificationCenter.default.removeObserver(self, name: .UIKeyboardWillChangeFrame, object: nil)
-        NotificationCenter.default.removeObserver(self, name: .UITextViewTextDidBeginEditing, object: nil)
-        NotificationCenter.default.removeObserver(self, name: .UIDeviceOrientationDidChange, object: nil)
-    }
-
-    @objc
-    func handleTextViewDidBeginEditing(_ notification: Notification) {
-        if scrollsToBottomOnKeybordBeginsEditing {
-            guard let inputTextView = notification.object as? InputTextView, inputTextView === messageInputBar.inputTextView else { return }
-            messagesCollectionView.scrollToBottom(animated: true)
-        }
-    }
-
-    @objc
-    func handleKeyboardDidChangeState(_ notification: Notification) {
-        guard let keyboardEndFrame = notification.userInfo?[UIKeyboardFrameEndUserInfoKey] as? CGRect else { return }
-
-        if (keyboardEndFrame.origin.y + keyboardEndFrame.size.height) > UIScreen.main.bounds.height {
-            // Hardware keyboard is found
-            messageCollectionViewBottomInset = view.frame.size.height - keyboardEndFrame.origin.y - iPhoneXBottomInset
-        } else {
-            //Software keyboard is found
-            let afterBottomInset = keyboardEndFrame.height > keyboardOffsetFrame.height ? (keyboardEndFrame.height - iPhoneXBottomInset) : keyboardOffsetFrame.height
-            let differenceOfBottomInset = afterBottomInset - messageCollectionViewBottomInset
-            let contentOffset = CGPoint(x: messagesCollectionView.contentOffset.x, y: messagesCollectionView.contentOffset.y + differenceOfBottomInset)
-
-            if maintainPositionOnKeyboardFrameChanged {
-                messagesCollectionView.setContentOffset(contentOffset, animated: false)
-            }
-
-            messageCollectionViewBottomInset = afterBottomInset
-        }
-    }
-    
-    fileprivate var keyboardOffsetFrame: CGRect {
-        guard let inputFrame = inputAccessoryView?.frame else { return .zero }
-        return CGRect(origin: inputFrame.origin, size: CGSize(width: inputFrame.width, height: inputFrame.height - iPhoneXBottomInset))
-    }
-    
-    /// On the iPhone X the inputAccessoryView is anchored to the layoutMarginesGuide.bottom anchor so the frame of the inputAccessoryView
-    /// is larger than the required offset for the MessagesCollectionView
-    ///
-    /// - Returns: The safeAreaInsets.bottom if its an iPhoneX, else 0
-    fileprivate var iPhoneXBottomInset: CGFloat {
-        if #available(iOS 11.0, *) {
-            guard UIScreen.main.nativeBounds.height == 2436 else { return 0 }
-            return view.safeAreaInsets.bottom
-        }
-        return 0
-    }
-}
-
-// MARK: - Menu Handling
-extension MessagesViewController {
-    private var navigationBarFrame: CGRect {
-        guard let navigationController = navigationController, !navigationController.navigationBar.isHidden else {
-            return .zero
-        }
-        return navigationController.navigationBar.frame
-    }
-    
-    /// Add observer for `UIMenuControllerWillShowMenu` notification
-    fileprivate func addMenuControllerObservers() {
-        NotificationCenter.default.addObserver(self, selector: #selector(MessagesViewController.menuControllerWillShow(notification:)), name: .UIMenuControllerWillShowMenu, object: nil)
-    }
-    
-    /// Remove observer for `UIMenuControllerWillShowMenu` notification
-    fileprivate func removeMenuControllerObservers() {
-        NotificationCenter.default.removeObserver(self, name: .UIMenuControllerWillShowMenu, object: nil)
-    }
-    
-    /// Show menuController and set target rect to selected bubble
-    @objc func menuControllerWillShow(notification: Notification) {
-        if let currentMenuController = notification.object as? UIMenuController,
-            let selectedIndexPath = selectedIndexPathForMenu {
-            NotificationCenter.default.removeObserver(self, name: .UIMenuControllerWillShowMenu, object: nil)
-            defer {
-                NotificationCenter.default.addObserver(self, selector: #selector(MessagesViewController.menuControllerWillShow(notification:)), name: .UIMenuControllerWillShowMenu, object: nil)
-                selectedIndexPathForMenu = nil
-            }
-            
-            currentMenuController.setMenuVisible(false, animated: false)
-            
-            guard let selectedCell = messagesCollectionView.cellForItem(at: selectedIndexPath) as? MessageCollectionViewCell else { return }
-            let selectedCellMessageBubbleFrame = selectedCell.convert(selectedCell.messageContainerView.frame, to: view)
-            
-            var messageInputBarFrame: CGRect = .zero
-            if let messageInputBarSuperview = messageInputBar.superview {
-                messageInputBarFrame = view.convert(messageInputBar.frame, from: messageInputBarSuperview)
-            }
-            
-            var topNavigationBarFrame: CGRect = navigationBarFrame
-            if navigationBarFrame != .zero, let navigationBarSuperview = navigationController?.navigationBar.superview {
-                topNavigationBarFrame = view.convert(navigationController!.navigationBar.frame, from: navigationBarSuperview)
-            }
-            
-            let menuHeight = currentMenuController.menuFrame.height
-            
-            let selectedCellMessageBubblePlusMenuFrame = CGRect(selectedCellMessageBubbleFrame.origin.x, selectedCellMessageBubbleFrame.origin.y - menuHeight, selectedCellMessageBubbleFrame.size.width, selectedCellMessageBubbleFrame.size.height + 2 * menuHeight)
-            
-            var targetRect: CGRect = selectedCellMessageBubbleFrame
-            currentMenuController.arrowDirection = .default
-            
-            /// Message bubble intersects with navigationBar and keyboard
-            if selectedCellMessageBubblePlusMenuFrame.intersects(topNavigationBarFrame) && selectedCellMessageBubblePlusMenuFrame.intersects(messageInputBarFrame) {
-                let centerY = (selectedCellMessageBubblePlusMenuFrame.intersection(messageInputBarFrame).minY + selectedCellMessageBubblePlusMenuFrame.intersection(topNavigationBarFrame).maxY) / 2
-                targetRect = CGRect(selectedCellMessageBubblePlusMenuFrame.midX, centerY, 1, 1)
-            } /// Message bubble only intersects with navigationBar
-            else if selectedCellMessageBubblePlusMenuFrame.intersects(topNavigationBarFrame) {
-                currentMenuController.arrowDirection = .up
-            }
-            
-            currentMenuController.setTargetRect(targetRect, in: view)
-            currentMenuController.setMenuVisible(true, animated: true)
-        }
     }
 }
